+++ backend/source/scrcpy/server/src/main/java/com/genymobile/scrcpy/Options.java
@@ -90,6 +90,7 @@ public class Options {
     private boolean listCameras;
     private boolean listCameraSizes;
     private boolean listApps;
+    private boolean listAllApps;
 
     // Options not used by the scrcpy client, but useful to use scrcpy-server directly
     private boolean sendDeviceMeta = true; // send device name and size
@@ -298,7 +299,7 @@ public class Options {
     }
 
     public boolean getList() {
-        return listEncoders || listDisplays || listCameras || listCameraSizes || listApps;
+        return listEncoders || listDisplays || listCameras || listCameraSizes || listApps || listAllApps;
     }
 
     public boolean getListEncoders() {
@@ -319,6 +320,10 @@ public class Options {
 
     public boolean getListApps() {
         return listApps;
+    }
+
+    public boolean getListAllApps() {
+        return listAllApps;
     }
 
     public boolean getSendDeviceMeta() {
@@ -729,6 +734,9 @@ public class Options {
                     break;
                 case "list_apps":
                     options.listApps = Boolean.parseBoolean(value);
+                    break;
+                case "list_all_apps":
+                    options.listAllApps = Boolean.parseBoolean(value);
                     break;
                 case "camera_id":
                     if (!value.isEmpty()) {